feat: Refactor SMTP configuration and service integration for improved email handling
